Backend/requirements.txt
requests==2.23.0
djongo==1.3.4
Django==3.1.7
linkedin_jobs_scraper==1.7.0
djangorestframework==3.12.3
django-cors-headers==3.7.0
beautifulsoup4==4.9.3
psycopg2==2.8.6
djangorestframework-jwt=1.11.0
pandas==1.2.3
